Refactoring artifact CLI contd
Move a part containing commands skeleton to separate file.
Fix parameters parsing for artifact write command.

Changelog: None
