+++ cli/mender-artifact/artifacts.go
@@ -25,7 +25,6 @@ import (
 	"os/exec"
 	"path/filepath"
 	"regexp"
-	"strings"
 
 	"github.com/mendersoftware/mender-artifact/areader"
 	"github.com/mendersoftware/mender-artifact/artifact"
@@ -840,189 +839,4 @@ func modifyArtifact(c *cli.Context) error {
 		}
 	}
 	return nil
-}
-
-func run() error {
-	app := cli.NewApp()
-	app.Name = "mender-artifact"
-	app.Usage = "Mender artifact read/writer"
-	app.UsageText = "mender-artifact [--version][--help] <command> [<args>]"
-	app.Version = Version
-
-	app.Author = "mender.io"
-	app.Email = "contact@mender.io"
-
-	//
-	// write
-	//
-	writeRootfs := cli.Command{
-		Name: "rootfs-image",
-		Action: func(c *cli.Context) error {
-			if len(c.StringSlice("device-type")) == 0 ||
-				len(c.String("artifact-name")) == 0 ||
-				len(c.String("update")) == 0 {
-				return cli.NewExitError("must provide `device-type`, `artifact-name` and `update`", 1)
-			}
-			if len(strings.Fields(c.String("artifact-name"))) > 1 { // check for whitespace in artifact-name
-				return cli.NewExitError("whitespace is not allowed in the artifact-name", 1)
-			}
-			return writeArtifact(c)
-		},
-	}
-
-	writeRootfs.Flags = []cli.Flag{
-		cli.StringFlag{
-			Name:  "update, u",
-			Usage: "Update `FILE`.",
-		},
-		cli.StringSliceFlag{
-			Name: "device-type, t",
-			Usage: "Type of device(s) supported by the update. You can specify multiple " +
-				"compatible devices providing this parameter multiple times.",
-		},
-		cli.StringFlag{
-			Name:  "artifact-name, n",
-			Usage: "Name of the artifact",
-		},
-		cli.StringFlag{
-			Name:  "output-path, o",
-			Usage: "Full path to output artifact file.",
-		},
-		cli.IntFlag{
-			Name:  "version, v",
-			Usage: "Version of the artifact.",
-			Value: LatestFormatVersion,
-		},
-		cli.StringFlag{
-			Name:  "key, k",
-			Usage: "Full path to the private key that will be used to sign the artifact.",
-		},
-		cli.StringSliceFlag{
-			Name: "script, s",
-			Usage: "Full path to the state script(s). You can specify multiple " +
-				"scripts providing this parameter multiple times.",
-		},
-	}
-
-	write := cli.Command{
-		Name:  "write",
-		Usage: "Writes artifact file.",
-		Subcommands: []cli.Command{
-			writeRootfs,
-		},
-	}
-
-	key := cli.StringFlag{
-		Name: "key, k",
-		Usage: "Full path to the public key that will be used to verify " +
-			"the artifact signature.",
-	}
-
-	//
-	// validate
-	//
-	validate := cli.Command{
-		Name:        "validate",
-		Usage:       "Validates artifact file.",
-		Action:      validateArtifact,
-		UsageText:   "mender-artifact validate [options] <pathspec>",
-		Description: "This command validates artifact file provided by pathspec.",
-	}
-	validate.Flags = []cli.Flag{
-		key,
-	}
-
-	//
-	// read
-	//
-	read := cli.Command{
-		Name:        "read",
-		Usage:       "Reads artifact file.",
-		Action:      readArtifact,
-		UsageText:   "mender-artifact read [options] <pathspec>",
-		Description: "This command validates artifact file provided by pathspec.",
-	}
-
-	read.Flags = []cli.Flag{
-		key,
-	}
-
-	//
-	// sign
-	//
-	sign := cli.Command{
-
-		Name:        "sign",
-		Usage:       "Signs existing artifact file.",
-		Action:      signExisting,
-		UsageText:   "mender-artifact sign [options] <pathspec>",
-		Description: "This command signs artifact file provided by pathspec.",
-	}
-	sign.Flags = []cli.Flag{
-		key,
-		cli.StringFlag{
-			Name: "output-path, o",
-			Usage: "Full path to output signed artifact file; " +
-				"if none is provided existing artifact will be replaced with signed one",
-		},
-		cli.BoolFlag{
-			Name:  "force, f",
-			Usage: "Force creating new signature if the artifact is already signed",
-		},
-	}
-
-	//
-	// modify existing
-	//
-	modify := cli.Command{
-		Name:        "modify",
-		Usage:       "Modifies image or artifact file.",
-		Action:      modifyArtifact,
-		UsageText:   "mender-artifact modify [options] <pathspec>",
-		Description: "This command modifies existing image or artifact file provided by pathspec.",
-	}
-
-	modify.Flags = []cli.Flag{
-		cli.StringFlag{
-			Name:  "key, k",
-			Usage: "Full path to the private key that will be used to sign the artifact after modifying.",
-		},
-		cli.StringFlag{
-			Name:  "server-uri, u",
-			Usage: "Mender server URI; the default URI will be replaced with given one.",
-		},
-		cli.StringFlag{
-			Name: "server-cert, c",
-			Usage: "Full path to the certificate file that will be used for validating " +
-				"Mender server by the client.",
-		},
-		cli.StringFlag{
-			Name: "verification-key, v",
-			Usage: "Full path to the public verification key that is used by the client  " +
-				"to verify the artifact.",
-		},
-		cli.StringFlag{
-			Name:  "name, n",
-			Usage: "New name of the artifact.",
-		},
-		cli.StringFlag{
-			Name:  "tenant-token, t",
-			Usage: "Full path to the tenant token that will be injected into modified file.",
-		},
-	}
-
-	app.Commands = []cli.Command{
-		write,
-		read,
-		validate,
-		sign,
-		modify,
-	}
-	return app.Run(os.Args)
-}
-
-func main() {
-	if err := run(); err != nil {
-		os.Exit(1)
-	}
 }
